+++ packages/elements/src/internals/machines/sign-in/types/verification.types.ts
@@ -19,12 +19,16 @@ export type SignInVerificationTags =
 export type SignInVerificationSubmitEvent = { type: 'SUBMIT' };
 export type SignInVerificationFactorUpdateEvent = { type: 'STRATEGY.UPDATE'; factor: SignInFactor | undefined };
 export type SignInVerificationRetryEvent = { type: 'RETRY' };
+export type SignInVerificationStrategyRegisterEvent = { type: 'STRATEGY.REGISTER'; factor: SignInFactor };
+export type SignInVerificationStrategyUnregisterEvent = { type: 'STRATEGY.UNREGISTER'; factor: SignInFactor };
 
 export type SignInVerificationEvents =
   | ErrorActorEvent
   | SignInVerificationSubmitEvent
   | SignInVerificationFactorUpdateEvent
   | SignInVerificationRetryEvent;
+  | SignInVerificationStrategyRegisterEvent
+  | SignInVerificationStrategyUnregisterEvent;
 
 // ---------------------------------- Input ---------------------------------- //
 
@@ -41,6 +45,7 @@ export interface SignInVerificationContext {
   formRef: ActorRefFrom<typeof FormMachine>;
   parent: ActorRefFrom<TSignInRouterMachine>;
   loadingStep: 'verifications';
+  registeredStrategies: Set<SignInFactor>;
   resendable: boolean;
   resendableAfter: number;
 }